♻️ Move code from cli_hooks to respective modules
