+++ src/cli.rs
@@ -1,7 +1,6 @@
 use crate::vpn::util::ConnectionProtocol;
 use structopt::StructOpt;
 
-pub mod cli_hooks;
 pub mod configure;
 pub mod initialize;
